Move app creation to separate module
