+++ portal/factories/app.py
@@ -9,36 +9,36 @@ from flask import Flask
 import redis
 from werkzeug.contrib.profiler import ProfilerMiddleware
 
-from .audit import configure_audit_log
-from .config import DefaultConfig
-from .csrf import csrf, csrf_blueprint
-from .database import db
-from .dogpile import dogpile_cache
-from .extensions import authomatic, recaptcha
-from .extensions import babel, celery, mail, oauth, session, user_manager
-from .models.app_text import app_text
-from .models.coredata import configure_coredata
-from .models.role import ROLE
-from .views.assessment_engine import assessment_engine_api
-from .views.audit import audit_api
-from .views.auth import auth, capture_next_view_function
-from .views.coredata import coredata_api
-from .views.clinical import clinical_api
-from .views.demographics import demographics_api
-from .views.extend_flask_user import reset_password_view_function
-from .views.fhir import fhir_api
-from .views.filters import filters_blueprint
-from .views.group import group_api
-from .views.intervention import intervention_api
-from .views.patient import patient_api
-from .views.patients import patients
-from .views.procedure import procedure_api
-from .views.portal import portal, page_not_found, server_error
-from .views.organization import org_api
-from .views.scheduled_job import scheduled_job_api
-from .views.tou import tou_api
-from .views.truenth import truenth_api
-from .views.user import user_api
+from ..audit import configure_audit_log
+from ..config import DefaultConfig
+from ..csrf import csrf, csrf_blueprint
+from ..database import db
+from ..dogpile import dogpile_cache
+from ..extensions import authomatic, recaptcha
+from ..extensions import babel, mail, oauth, session, user_manager
+from ..models.app_text import app_text
+from ..models.coredata import configure_coredata
+from ..models.role import ROLE
+from ..views.assessment_engine import assessment_engine_api
+from ..views.audit import audit_api
+from ..views.auth import auth, capture_next_view_function
+from ..views.coredata import coredata_api
+from ..views.clinical import clinical_api
+from ..views.demographics import demographics_api
+from ..views.extend_flask_user import reset_password_view_function
+from ..views.fhir import fhir_api
+from ..views.filters import filters_blueprint
+from ..views.group import group_api
+from ..views.intervention import intervention_api
+from ..views.patient import patient_api
+from ..views.patients import patients
+from ..views.procedure import procedure_api
+from ..views.portal import portal, page_not_found, server_error
+from ..views.organization import org_api
+from ..views.scheduled_job import scheduled_job_api
+from ..views.tou import tou_api
+from ..views.truenth import truenth_api
+from ..views.user import user_api
 
 SITE_CFG = 'site.cfg'
 DEFAULT_BLUEPRINTS = (
@@ -172,7 +172,7 @@ def configure_extensions(app):
     session.init_app(app)
 
     # celery - task queue for asynchronous tasks
-    celery.init_app(app)
+    # celery.init_app(app)
 
     # babel - i18n
     babel.init_app(app)
